test(server-core): colocate game reconnect guard coverage
